Added new route to add art work and refactored add wrok form to new component
