+++ client/src/pages/addWork.jsx
@@ -1,0 +1,9 @@
+import AddArtwork from "../components/AddArtwork";
+
+export default function AddWork(props) {
+    return (
+        <>
+        <AddArtwork/>
+        </>
+    )
+}
